Move ancestry logic elsewhere
