+++ lib/nanoc-conref-fs.rb
@@ -1,3 +1,4 @@
+require 'nanoc-conref-fs/ancestry'
 require 'nanoc-conref-fs/conref-fs'
 require 'nanoc-conref-fs/conrefifier'
 require 'nanoc-conref-fs/datafiles'
